quality of life tweaks and unsolvable puzzle bug
Looks great so far. Here are two changes I'd like to make:

- Make the first guess completely at random. This adds some variety by not
  always generating an `xxyy` pattern of guess.
- Have the drop down inputs start blank. A player can add in colors as they
  figure them out and the blanks help show what's left to deduce.

Additionally, I think there is a bug. I was presented with this game:

    ppoo !.
    pprg .
    pygo .
    ppyb
    pgrr !.

Where a `!` represents a black peg and a `.` represents a white one. My
deductions led me to believe the solution could have been *either* `oror` or
`roor`. There is nothing to disambiguate the order of `o` and `r` in the first
two slots.