suggestion for randomness fix
I saw this condition in `findBestGuess`:

  score > bestScore || (score === bestScore && Math.random() < 0.5)

I think this random choice gives the advantage to items later in the
iteration. To implement stream sampling, we should choose this element
with 1/n probability, where `n` is the number of items seen. As a bonus,
I think you can simply reset the counter to 1 when `score > bestScore`
such that it chooses it with 100% probability.