share link tweaks
I have some proposals to tweak the serialization format:

- Since the total input is just `secret` + a series of inputs, you don't
  need to add the number of guesses to the string
- If we're just stringifying each digit, there's no reason to convert to
  a digit in the first place.
- That said, I think we should give the bitshifting approach another try
  now that we are just storing slots with values 1-6.
- Instead of base64, perhaps just `toString(36)` would work well enough?

...

Is it necessary to split the code into segments and join with '-'? The
number of inputs should be incredibly small (2-5) so I don't think a
single number is going to get problematic.

...

Great, the last thing I would suggest is adding a prefix of "1." as a
way of versioning in case we want to change the serialization later