advanced input mode
I think something I would like most is something more than a dropdown
box as I'm trying to solve a puzzle. For example, I might realize that
there's a yellow in the first second or third slots, but it's hard to
mark that without simply picking yellow, which would be wrong. I don't
have a great idea for what the UI might look like. A multiselect form
which only validates when one item is selected would be a crude way of
getting the result, but maybe you can come up with something a little
nicer. Maybe it's also behind a toggle for advanced input.

...

I like that it shows the input when I check something, but if I select
more than one thing, I think it should revert to hiding again,
especially so that the form is invalidated.

...

Let's also remove the "Position 1" text in each box and stack all 6
entries vertically in a single column per slot

...

I like the "?" placeholder! Can we show that when nothing is selected
too? I also think removing the min-width or making them even more narrow
is fine.

...

And maybe we can move the "Switch to advanced mode" button to be
right-justified as well? It could either flow inline with the input or
be pushed to the line after if it doesn't fit

...

I think you need to put it in the same container as the input buttons
themselves, otherwise it will push them down

...

Let's have the button align itself to the right in the container, and
also the padding/margin appears to be lost in the advanced mode

...

Oh, I meant the bottom margin. I'm still not seeing it on
`.player-input-advanced`