warn if deserialization leads to an ambiguous puzzle
(Note: I don't think this actually happened, but I do know that I am the
kind of person that might troll someone by hand-crafting an ambiguous
puzzle link.)

The serializer currently takes a secret and each guess and packs them
into a string together. This works fine, but it also means that someone
could construct a serialized payload that didn't represent a valid
puzzle -- e.g., something where the number of remaining possibilities
was more than 1 leading to an ambiguous solution. The first thing I
would like to do is validate that the puzzle has a unique solution which
matches the encoded secret and warn the user otherwise. See
`filterPossibleSolutions` for the functionality you need.

...

I think it's okay to load the puzzle as it was interpreted without
generating a new one so long as the warning is shown

...

I don't think `possibleSolutions.length` can be 0, and if it is 1, I
think it necessarily has to be the secret. I claim these are both true
because the feedback is not stored in the serialized format, it is
computed based on the serialized format, so it is impossible to
eliminate the secret accidentally. We can leave these branches in, but
they should be assertions instead of handled errors.